Move unix platforms into the same folder to move the unix files into it as well
