+++ src/unix/apple/macos/component/x86.rs
@@ -77,7 +77,7 @@ impl Components {
     }
 }
 
-#[doc = include_str!("../../../../md_doc/component.md")]
+#[doc = include_str!("../../../../../md_doc/component.md")]
 pub struct Component {
     temperature: f32,
     max: f32,
