+++ src/unix/apple/macos/process.rs
@@ -14,7 +14,7 @@ use crate::{DiskUsage, Gid, Pid, ProcessExt, ProcessRefreshKind, ProcessStatus, 
 use crate::sys::process::ThreadStatus;
 use crate::sys::system::Wrap;
 
-#[doc = include_str!("../../../md_doc/process.md")]
+#[doc = include_str!("../../../../md_doc/process.md")]
 pub struct Process {
     pub(crate) name: String,
     pub(crate) cmd: Vec<String>,
@@ -316,7 +316,7 @@ fn check_if_pid_is_alive(pid: Pid, check_if_alive: bool) -> bool {
             return true;
         }
         // `kill` failed but it might not be because the process is dead.
-        let errno = crate::libc_errno();
+        let errno = crate::unix::libc_errno();
         // If errno is equal to ESCHR, it means the process is dead.
         !errno.is_null() && *errno != libc::ESRCH
     }
