+++ src/unix/apple/macos/utils.rs
