+++ src/unix/apple/network.rs
@@ -164,7 +164,7 @@ impl NetworksExt for Networks {
     }
 }
 
-#[doc = include_str!("../../md_doc/network_data.md")]
+#[doc = include_str!("../../../md_doc/network_data.md")]
 #[derive(PartialEq, Eq)]
 pub struct NetworkData {
     current_in: u64,
