+++ src/unix/apple/system.rs
@@ -74,7 +74,7 @@ declare_signals! {
     _ => None,
 }
 
-#[doc = include_str!("../../md_doc/system.md")]
+#[doc = include_str!("../../../md_doc/system.md")]
 pub struct System {
     process_list: HashMap<Pid, Process>,
     mem_total: u64,
@@ -327,7 +327,7 @@ impl SystemExt for System {
     }
 
     fn refresh_users_list(&mut self) {
-        self.users = crate::apple::users::get_users_list();
+        self.users = crate::sys::users::get_users_list();
     }
 
     // COMMON PART
