+++ src/unix/linux/network.rs
@@ -126,7 +126,7 @@ impl NetworksExt for Networks {
     }
 }
 
-#[doc = include_str!("../../md_doc/network_data.md")]
+#[doc = include_str!("../../../md_doc/network_data.md")]
 pub struct NetworkData {
     /// Total number of bytes received over interface.
     rx_bytes: u64,
