+++ src/windows/network_helper.rs
